v.1.1.7
Improved work with forms - now in case of a successful response there will be redirects and dynamic updating of information. Added containerization for bug checking. fixed a stupid mistake with serializers. Added interaction with SSE - on the client side it does not need changes, on the server side it is implemented more incorrectly than correctly(it works, but is very primitive)