v.1.2.6
CompanyCharacteristics table has been removed since I decided not to give the opportunity to upgrade the company (this is a simulation after all),
so I moved the important attributes - production speed and volume to the CompanyType table to make it easier to influence
on certain types of companies under specific events.
Now the warehouse for the company's target products has a volume of 50_000 units, for the rest it will be 10_000. Events added.